task3_incremental_processing/requirements.txt
greenlet==3.2.4
joblib==1.5.2
kafka-python==2.2.15
numpy==2.3.3
pandas==2.3.3
py4j==0.10.9.7
pyspark==3.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
SQLAlchemy==2.0.44
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
